Edits to Physical Layer
A heavier hand was taken in moving items from the physical layer to the
data link layer.  The physical layer wording has been revised to
(hopefully) minimize or eliminate references to the data link layer.

One exception is in the text of sync burst.  It is mentioned that the
data link layer may utilize different sync burst codings to identify the
type of data to follow.