Updates what consitutes a transmission, fixes data reference, adds CSMA
A transmission can be composed of multiple Sync Burst + Data intervals.
The wording now makes that clear and a graphic is added to help.

Data bits do not have to be in byte multiples.  There are cases where
higher layers use extra bits to flush encoders, etc.  The Physical Layer
will not define the number or organization of the data bits.  However,
they do need to be in pairs for THIS physical layer, since we're using
4FSK (2 bits per symbol).

The Randomizer section was rewritten to deal with data as bits and not
bytes.

CSMA was moved from the Data Link layer to the Physical layer, and
worded in a generic form so that higher layers can set the parameters
and determine the cases when CSMA is required.