Extensive changes to fold in ECC/FEC information across frame types
This is a major update to document the ECC/FEC processing flows.

Specific details of the various elements have been moved to sections in the Appendix, and stated succinctly in the main body of the spec.

Some work on the application layer has been started (work in progress).

The Data Link Layer should be in relatively good shape now.